+++ modules/core/petra/petra-url-pattern-mapper/src/main/java/com/liferay/petra/url/pattern/mapper/internal/BaseURLPatternMapper.java
@@ -12,7 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.remote.cors.internal.url.pattern.mapper;
+package com.liferay.petra.url.pattern.mapper.internal;
+
+import com.liferay.petra.url.pattern.mapper.URLPatternMapper;
 
 import java.net.URI;
 import java.net.URISyntaxException;
